+++ src/types/components/auth-form.ts
@@ -1,5 +1,5 @@
 import { UseFormGetValues, UseFormRegister, UseFormTrigger } from 'react-hook-form';
-import { SignupFormData } from '@/lib/utils/authValidate';
+import { SignupFormData } from '@/validation/authValidate';
 
 export interface AuthInputProps {
   register: UseFormRegister<SignupFormData>;
